feat: start and stop local networks
Provide administration for local test networks during workflow runs, which we were repeating over
multiple jobs.

The `start` action brings up a local network, by default using the latest version of the testnet
binary, which it pulls from S3. There is a `build` input to instruct it to build the binary from
source if necessary.

The `stop` action kills the nodes, then tars and uploads the log files.

We can extend it for our workflows that use heaptrack.