feat: use the node manager to launch testnet
The `testnet` binary has been replaced with the node manager. We use the node manager's `run`,
`join` and `kill` commands for managing local networks.

As with the `testnet` binary, we've retained the possibility of building the node manager from
source, which is useful for testing purposes. However, since the node manager is more decoupled from
`safe_network`, changes should probably be infrequent. Most of the time we can just download the
latest version of the node manager.